Refactoring of GenC: split implementation in several smaller files
